+++ package.json
@@ -68,6 +68,7 @@
     "fake-indexeddb": "^6.2.5",
     "http-server": "^14.1.1",
     "jest": "^29.7.0",
+    "jest-dev-server": "^11.0.0",
     "jest-environment-jsdom": "^29.7.0",
     "jest-fixed-jsdom": "^0.0.9",
     "jest-image-snapshot": "^6.4.0",
