+++ test/visual-global-teardown.ts
@@ -1,0 +1,6 @@
+import { teardown as teardownDevServer } from 'jest-dev-server'
+
+module.exports = async function globalTeardown() {
+  await teardownDevServer(globalThis.servers)
+  // Your global teardown
+}
